created stub methods for characters dao unit test
